+++ src/index.js
@@ -3,8 +3,8 @@ import MJMLElementsCollection, {registerElement} from './MJMLElementsCollection'
 module.exports = {
   elements: MJMLElementsCollection,
   documentParser: require('./documentParser'),
-  MJMLElement: require('./components/MJMLElement'),
-  MJMLColumnElement: require('./components/MJMLColumnElement'),
+  MJMLElement: require('./components/decorators/MJMLElement'),
+  MJMLColumnElement: require('./components/decorators/MJMLColumnElement'),
   mjml2html: require('./mjml2html').default,
   registerElement: registerElement,
   version: require('../package.json').version
